test: single test_smiles_datasets.py, also testing base methods.
